+++ src/index.js
@@ -2,7 +2,7 @@ import './style.css';
 import displayTasks from './display-tasks';
 import {
   addTask, deleteTask, submit, tasksContainer,
-} from './add-and-delete';
+} from './class';
 // import dragFeature from './dragAndDrop.js';
 
 /* Add and remove */
